Atualização do grupo J e instruções de reprodução

+++ benchmark/benchmark_com_assertions_com_alinhamento_explicito_v4/C_uint64_t_misaligned.c
@@ -3,7 +3,7 @@
 #include "verifier.h"
 /* Grupo C — vetor desalinhado | esperado: VIOLATION */
 int main(void) {
-    _Alignas(uint32_t)  uint64_t vec[8];
+    _Alignas(uint64_t)  uint64_t vec[8];
     uint64_t *p = (uint64_t*)((char*)vec + 1);
     uint64_t z = *p;
     (void)z;
